requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
docker==7.1.0
ecdsa==0.19.1
eventkit==1.0.3
exchange_calendars==4.10
fastapi==0.115.12
finnhub-python==2.4.23
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
ib-insync==0.9.86
idna==3.10
korean-lunar-calendar==0.3.1
nest-asyncio==1.6.0
numpy==2.2.4
pandas==2.2.3
pandas_market_calendars==5.0.0
pandas_ta==0.3.14b0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
pyluach==2.2.0
pyreadline3==3.5.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
schedule==1.2.2
security==1.3.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
toolz==1.0.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.0
email-validator>=2.0
httpx>=0.24
psutil>=5.9